refactor: take out necessary requests
